requirements.txt
Django==1.8.6
djangorestframework==3.3.1
pep8==1.6.2
wheel==0.24.0
